Move Processor base class to core
